AimTux: Enable -Wall and fix warnings

+++ src/netvarmanager.cpp
@@ -24,7 +24,7 @@ RecvTable* NetVarManager::getTable(std::vector<RecvTable *> tables, const char *
 	if (tables.empty())
 		return NULL;
 
-	for (int i = 0; i < tables.size(); i++)
+	for (size_t i = 0; i < tables.size(); i++)
 	{
 		RecvTable *table = tables[i];
